SkypeWeb: - history sync (patch from MikalaiR) - refactored message sending
git-svn-id: http://svn.miranda-ng.org/main/trunk@12599 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/SkypeWeb/src/skype_contacts.cpp
@@ -216,7 +216,7 @@ void CSkypeProto::LoadContactList(const NETLIBHTTPREQUEST *response)
 	{
 		PushRequest(new GetContactsInfoRequest(token, skypenames), &CSkypeProto::LoadContactsInfo);
 
-		for (size_t i = 0; i < skypenames.getCount(); i++)
+		for (int i = 0; i < skypenames.getCount(); i++)
 			mir_free(skypenames[i]);
 		skypenames.destroy();
 	}
